refactor: move mocks to distinct files in mocks folder
