+++ src/agent/communicator/tests/mocks/mock_http_client.hpp
@@ -1,0 +1,38 @@
+#include <gmock/gmock.h>
+
+#include <ihttp_client.hpp>
+
+class MockHttpClient : public http_client::IHttpClient
+{
+public:
+    MOCK_METHOD(boost::beast::http::request<boost::beast::http::string_body>,
+                CreateHttpRequest,
+                (const http_client::HttpRequestParams& params),
+                (override));
+
+    MOCK_METHOD(boost::asio::awaitable<void>,
+                Co_PerformHttpRequest,
+                (const std::string& token,
+                 http_client::HttpRequestParams params,
+                 std::function<boost::asio::awaitable<std::string>()> messageGetter,
+                 std::function<void()> onUnauthorized,
+                 std::function<void(const std::string&)> onSuccess,
+                 std::function<bool()> loopRequestCondition),
+                (override));
+
+    MOCK_METHOD(boost::beast::http::response<boost::beast::http::dynamic_body>,
+                PerformHttpRequest,
+                (const http_client::HttpRequestParams& params),
+                (override));
+
+    MOCK_METHOD(std::optional<std::string>,
+                AuthenticateWithUuidAndKey,
+                (const std::string& host, const std::string& port, const std::string& uuid, const std::string& key),
+                (override));
+
+    MOCK_METHOD(
+        std::optional<std::string>,
+        AuthenticateWithUserPassword,
+        (const std::string& host, const std::string& port, const std::string& user, const std::string& password),
+        (override));
+};
